Update aero bios version to 16
Updating the bios version to 16, which includes below changes,
Updates to the ACPI table for the CAN bus controller (mcp251x) connected to
the Intel® Atom™ processor SPI bus, specifically configuration of the input
clock and interrupt.
Updates to allow the user to read and write the Intel® Atom™ processor GPIO
pins that are physically accessible through the 80-pin IO Expansion connector.